feat: CLI build can use JSON or Py input.
