+++ src/cycax/cli/cmd_build.py
@@ -5,6 +5,7 @@ from typing import Annotated
 
 import typer
 
+from cycax.cli.run import make_build_map
 from cycax.cycad.assembly_openscad import AssemblyOpenSCAD
 from cycax.cycad.engines.part_freecad import PartEngineFreeCAD
 from cycax.cycad.engines.part_openscad import PartEngineOpenSCAD
@@ -15,7 +16,8 @@ app = typer.Typer(context_settings={"help_option_names": ["-h", "--help"]}, no_a
 def load_cycax_json(filename: str) -> dict:
     _filename = Path(filename).expanduser().resolve().absolute()
     if not _filename.exists():
-        raise FileNotFoundError(f"File {_filename} does not exist")
+        msg = f"File {_filename} does not exist"
+        raise FileNotFoundError(msg)
     data = json.loads(_filename.read_text())
     return _filename, data
 
@@ -26,13 +28,15 @@ def build_freecad(
     filename: Annotated[str, typer.Argument(help="A JSON file with CyCAx Features")],
 ):
     logging.info("Listing objects")
-    _filename, data = load_cycax_json(filename)
-    if data.get("features"):
-        engine = PartEngineFreeCAD(name=data["name"], path=_filename.parent.parent)
-        engine._json_file = _filename
-        engine.build(None)
-    else:
-        logging.warning("No features found in JSON file")
+    build_order = make_build_map(filename)
+    for build in sorted(build_order.values(), key=lambda x: x["index"]):
+        _filename, data = load_cycax_json(build["path"])
+        if data.get("features"):
+            engine = PartEngineFreeCAD(name=data["name"], path=_filename.parent.parent)
+            engine._json_file = _filename
+            engine.build(None)
+        else:
+            logging.warning("No features found in JSON file")
 
 
 def build_part(engine, part_json_file: Path):
@@ -63,6 +67,6 @@ def build_openscad(
             _part_filename = _filename.parent / part["part_no"] / f"{part['part_no']}.json"
             build_part(PartEngineOpenSCAD, _part_filename)
             engine.add(part)
-        files = engine.build(_filename.parent)
+        engine.build(_filename.parent)
     else:
         logging.warning("No parts found in JSON file")
